requirements.txt
aniso8601==7.0.0
asgiref==3.2.10
Django==3.1.1
django-filter==2.4.0
django-graphql-jwt==0.3.1
djangorestframework==3.11.1
djangorestframework-simplejwt==4.4.0
graphene==2.1.8
graphene-django==2.13.0
graphene-file-upload==1.3.0
graphql-core==2.3.2
graphql-relay==2.0.1
promise==2.3
psycopg2==2.8.6
PyJWT==1.7.1
pytz==2020.1
Rx==1.6.1
singledispatch==3.4.0.3
six==1.15.0
sqlparse==0.3.1
Unidecode==1.1.1
